index restructuring
